abstracting event store settings
